+++ ChangeLog.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - Texture transform UV rotation: Using standard `_ST` property (Scale-Translation) by default. When rotation is enabled, scale values act as m00/m11 values of 2-by-2 rotation/scale matrix and are supplemented by two rotation values (for m01/m10).
 - Textures that fail to load don't cause the whole loading process to fail (thanks @Bersaelor for #117)
+- Unit Tests: Sample model list is now part of GltfSampleSet objects and not loaded from text file anymore
 ### Fixed
 - Removed shader compiler warnings for built-in shaders
 - Removed compiler warnings in Unity 2020.1/2020.2